Bug fix
It turns out the reason I was losing all my data from midnight to 3am
was because that data was already in the directory by the time I
checked it, however my first check of a new day was failing to account
for the fact that it was checking a new directory, comparing the new
file list length to the previous day's file list length, returning a
negative number. It thus decides that there are no new files and
doesn't download them, before finally setting the current directory
list length to the correct value.

Hopefully I've fixed it.